This PR optimizes the way `pendingExternalEvents` are handled by `DurableOrchestrationContext`. It reduced CPU usage by 50% and allocated memory by a lot. It replaces the stack and the generic reflection with a simpler construct that provides both, the `TaskCompletionSource<T>` and a way to retrieve the event type without obtaining the type of `tcs`

### Benchmarks

The following suite was used locally. I see no benchmarks in the repo and don't want to introduce a lot of noise in one PR.

```csharp
[Benchmark]
public Task OneEvent()
{
    var wait1 = ((IDurableOrchestrationContext)context).WaitForExternalEvent("test");
    context.RaiseEvent("test", null);
    return wait1;
} 

[Benchmark]
public Task TwoDifferentEvents()
{
    var wait1 = ((IDurableOrchestrationContext)context).WaitForExternalEvent("test");
    var wait2 = ((IDurableOrchestrationContext)context).WaitForExternalEvent("test2");
    
    context.RaiseEvent("test", null);
    context.RaiseEvent("test2", null);
    
    return Task.WhenAll(wait1, wait2);
}

[Benchmark]
public Task TwoSameEvents()
{
    var wait1 = ((IDurableOrchestrationContext)context).WaitForExternalEvent("test");
    var wait2 = ((IDurableOrchestrationContext)context).WaitForExternalEvent("test");

    context.RaiseEvent("test", null);
    context.RaiseEvent("test", null);

    return Task.WhenAll(wait1, wait2);
}
```

The results looks quite good for such a small change. Both CPU and allocations are greatly reduced

```init
BenchmarkDotNet=v0.13.1, OS=Windows 10.0.19042.1165 (20H2/October2020Update)
Intel Core i7-6700HQ CPU 2.60GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
.NET SDK=6.0.100-preview.7.21379.14
  [Host]     : .NET Core 3.1.18 (CoreCLR 4.700.21.35901, CoreFX 4.700.21.36305), X64 RyuJIT
  DefaultJob : .NET Core 3.1.18 (CoreCLR 4.700.21.35901, CoreFX 4.700.21.36305), X64 RyuJIT
```

|             Method |       Mean |    Error |   StdDev |  Gen 0 | Allocated |
|------------------- |-----------:|---------:|---------:|-------:|----------:|
|           OneEvent |   653.7 ns | 12.75 ns | 11.30 ns | 0.1345 |     424 B |
| **OneEvent (after)**| 256.4 ns |  5.10 ns |  8.52 ns | 0.0558 |     176 B |
| TwoDifferentEvents | 1,398.9 ns | 27.22 ns | 26.73 ns | 0.3223 |   1,016 B |
| **TwoDifferentEvents (after)** | 628.1 ns | 12.35 ns | 15.62 ns | 0.1650 |     520 B |
|      TwoSameEvents | 1,357.4 ns | 26.82 ns | 41.75 ns | 0.2766 |     872 B |
|      **TwoSameEvents (after)** | 641.7 ns | 12.37 ns | 11.57 ns | 0.1631 |     512 B |


### Pull request checklist

* [x] My changes **do not** require documentation changes
    * [ ] Otherwise: Documentation PR is ready to merge and referenced in `pending_docs.md`
* [ ] My changes **should not** be added to the release notes for the next release
    * [ ] Otherwise: I've added my notes to `release_notes.md`
* [x] My changes **do not** need to be backported to a previous version
    * [ ] Otherwise: Backport tracked by issue/PR #issue_or_pr
* [ ] I have added all required tests (Unit tests, E2E tests)
* [x] My changes **do not** require any extra work to be leveraged by OutOfProc SDKs
    * [ ] Otherwise: That work is being tracked here: #issue_or_pr_in_each_sdk