@cgillum 2021-09-09
Very cool! Thanks for contributing this. Give me a little time to study it and we'll look into merging it.