@Scooletz 2021-09-15
How is it possible @cgillum ? I've run locally V2 and it's ok?