@cgillum 2021-09-16
Looks like the Release build failed with this StyleCop failure:
> [error]src\WebJobs.Extensions.DurableTask\ContextImplementations\DurableOrchestrationContext.cs(817,87): Error SA1028: Code should not contain trailing whitespace